requirements.txt
docxtpl
jinja2
python-docx